+++ src/test/java/net/kaczmarzyk/spring/data/jpa/domain/LikeIgnoreCaseTest.java
@@ -51,14 +51,12 @@ public class LikeIgnoreCaseTest extends IntegrationTestBase {
     @Test
     public void filtersFirstLevelPropertyIgnoringCase() {
         LikeIgnoreCase<Customer> lastNameSimpson = new LikeIgnoreCase<>(queryCtx, "lastName", "sIMPSOn");
-        lastNameSimpson.setLocale(Locale.getDefault());
         List<Customer> result = customerRepo.findAll(lastNameSimpson);
         assertThat(result)
             .hasSize(2)
             .containsOnly(homerSimpson, margeSimpson);
         
         LikeIgnoreCase<Customer> firstNameWithO = new LikeIgnoreCase<>(queryCtx, "firstName", "o");
-        firstNameWithO.setLocale(Locale.getDefault());
         result = customerRepo.findAll(firstNameWithO);
         assertThat(result)
             .hasSize(2)
@@ -68,7 +66,6 @@ public class LikeIgnoreCaseTest extends IntegrationTestBase {
     @Test
     public void filtersByNestedPropertyIgnoringCase() {
         LikeIgnoreCase<Customer> streetWithEvergreen = new LikeIgnoreCase<>(queryCtx, "address.street", "EvErGReeN");
-        streetWithEvergreen.setLocale(Locale.getDefault());
         List<Customer> result = customerRepo.findAll(streetWithEvergreen);
         assertThat(result).hasSize(2).containsOnly(homerSimpson, margeSimpson);
     }