+++ src/test/java/net/kaczmarzyk/spring/data/jpa/domain/NotLikeIgnoreCaseLocaleIntegrationTest.java
@@ -22,7 +22,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
-import java.util.Locale;
+
 
 import static net.kaczmarzyk.spring.data.jpa.CustomerBuilder.customer;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -40,24 +40,12 @@ public class NotLikeIgnoreCaseLocaleIntegrationTest extends IntegrationTestBase 
     }
 
     @Test
-    public void usesLocaleWhenPerformingComparisons() {
-
-        // English locale
+    public void filtersIgnoringCaseAccordingToDbCollation() {
         NotLikeIgnoreCase<Customer> simpsons = new NotLikeIgnoreCase<>(queryCtx, "lastName", new String[] { "i" });
-        simpsons.setLocale(Locale.ENGLISH);
         List<Customer> simpsonsFound = customerRepo.findAll(simpsons);
 
         assertThat(simpsonsFound)
                 .hasSize(1)
                 .containsOnly(homerWithTurkishCapitalI);
-
-        // Turkish locale
-        simpsons = new NotLikeIgnoreCase<>(queryCtx, "lastName", new String[] { "i" });
-        simpsons.setLocale(new Locale("tr", "TR"));
-        simpsonsFound = customerRepo.findAll(simpsons);
-
-        assertThat(simpsonsFound)
-                .hasSize(2)
-                .containsOnly(homerWithLowercaseI, homerWithEnglishCapitalI); // homerWithLowercaseI is included because test db does not use Turkish collation
     }
 }